feat(S13P31A706-44): API 호출 처리

+++ backend/src/main/java/com/topoom/missingcase/dto/MissingCaseDto.java
@@ -26,7 +26,7 @@ public class MissingCaseDto {
         private Long id;
         private String personName;
         private String gender;
-        private Integer currentAge;
+        private Short currentAge;
         private String occurredLocation;
         private LocalDateTime occurredAt;
         private String progressStatus;
